refactor(peppermint): extract peppermint worker to separate file
